commit 2

+++ src/models/cliente.rs
@@ -1,5 +1,6 @@
+#[derive(Default)]
 pub struct Cliente {
-  pub id : i32,
+  pub id : usize,
   pub nome: String,
   pub cpf: String,
   pub endereco: String,
